improve comments ui tests
Fix the getBoardsByName api call

Add logger for getting board by name though api

Fix comments tests cases json
