+++ .env.example
@@ -2,3 +2,4 @@ EMAIL=
 PASSWORD=
 TRELLO_KEY=
 TRELLO_TOKEN=
+ORGANIZATION_ID=
