+++ tests/tests_UI/card_comments/test_add_comments.spec.js
@@ -1,0 +1,17 @@
+const { test, expect } = require('../../../fixtures/create_and_open_card.js');
+const path = require('path');
+const fs = require('fs');
+
+const testCasesPath = path.resolve(__dirname, '../../../data/add-comments-cases.json');
+const listTestCases = JSON.parse(fs.readFileSync(testCasesPath, 'utf-8') );
+
+const testToRun = listTestCases.find(tc => tc.id == 1);
+
+for (const testCase of listTestCases) {
+  test(`CM-${testCase.id} - ${testCase.title} @regression @positivo`, async ({ card }) => {
+    await card.addComment(testCase.comment);
+    const actualComments = await card.getCommentsTextOnly();
+    const actualText = actualComments.join('\n'); 
+    expect(actualText).toContain(testCase.comment);
+  });
+}
